% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe_data.R
\name{as_huxtable.t1_shape}
\alias{as_huxtable.t1_shape}
\title{Convert a \code{t1_summary} object to a \code{huxtable}}
\usage{
\method{as_huxtable}{t1_shape}(
  x,
  ...,
  font_size = getOption("tableone.font_size", 8),
  font = getOption("tableone.font", "Arial"),
  footer_text = NULL,
  show_binary_value = NULL
)
}
\arguments{
\item{x}{the \code{t1_summary} object as produced by describe_population}

\item{...}{not used}

\item{font_size}{(optional) the font size for the table in points}

\item{font}{(optional) the font family for the table (which will be matched to
closest on your system)}

\item{footer_text}{any text that needs to be added at the end of the table,
setting this to FALSE dsables the whole footer (as does
\code{options("tableone.hide_footer"=TRUE)}).}

\item{show_binary_value}{if set this will filter the display of covariates where the number of possibilities
is exactly 2 to this value.}
}
\value{
a formatted table as a \code{huxtable}
}
\description{
Convert a \code{t1_summary} object to a \code{huxtable}
}
