% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe_population.R
\name{as_huxtable.t1_summary}
\alias{as_huxtable.t1_summary}
\title{Convert a \code{t1_summary} object to a \code{huxtable}}
\usage{
\method{as_huxtable}{t1_summary}(
  x,
  ...,
  layout = "single",
  override_percent_dp = list(),
  override_real_dp = list(),
  font_size = getOption("tableone.font_size", 8),
  font = getOption("tableone.font", "Arial"),
  footer_text = NULL,
  show_binary_value = NULL
)
}
\arguments{
\item{x}{the \code{t1_summary} object as produced by describe_population}

\item{...}{not used}

\item{layout}{(optional) various layouts are defined as default. As of this
version of \code{tableone} they are
"relaxed","compact","micro","simple","single","missing". The layouts can be
customised using the options \verb{options("tableone.format_list"=list(...)")},
and this is described in more detail in the vignettes.}

\item{override_percent_dp}{(optional) a named list of overrides for the default
precision of formatting percentages, following a \verb{c(<colname_1> = 2, <colname_2> = 4, ...)} format. columns not present in this list
will use the defaults defined in the layout. See the vignette on customisation.}

\item{override_real_dp}{(optional) a named list of overrides for the default
precision of formatting real values, following a \verb{c(<colname_1> = 2, <colname_2> = 4, ...)} format. columns not present in this list
will use the defaults defined in the layout. See the
\code{utils::vignette("customisation", package="tableone")}.}

\item{font_size}{(optional) the font size for the table in points}

\item{font}{(optional) the font family for the table (which will be matched to
closest on your system)}

\item{footer_text}{any text that needs to be added at the end of the table,
setting this to FALSE dsables the whole footer (as does
\code{options("tableone.hide_footer"=TRUE)}).}

\item{show_binary_value}{if set this will filter the display of covariates where the number of possibilities
is exactly 2 to this value.}
}
\value{
a formatted table as a \code{huxtable}
}
\description{
Convert a \code{t1_summary} object to a \code{huxtable}
}
