% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_population.R
\name{as_t1_signif}
\alias{as_t1_signif}
\title{Compares the population against an intervention}
\usage{
as_t1_signif(
  df,
  ...,
  label_fn = label_extractor(df),
  units = extract_units(df),
  override_type = list(),
  override_method = list()
)
}
\arguments{
\item{df}{a dataframe of individual observations. If using the \code{tidyselect}
syntax data grouping defines the intervention group and should be present. if the
formula interface is used the first variable in the right hand side of the
formula is used as the intervention, in which case grouping is ignored.}

\item{...}{the columns of variables we wish to summarise. This can be given as
a \code{tidyselect} specification (see \code{utils::vignette("syntax", package = "tidyselect")}),
identifying the columns. Alternatively it can be given as a formula of the nature

\code{outcome ~ intervention + covariate_1 + covariate_2 + ...} .

which may be more convenient if you are going on to do a model fit later. If the
latter format the left hand side is ignored (outcomes are not usual in
this kind of table).}

\item{label_fn}{(optional) a function for mapping a co-variate column name to
printable label. This is by default a no-operation and the output table
will contain the dataframe column names as labels. A simple alternative
would be some form of \link[dplyr:case_when]{dplyr::case_when} lookup, or a string function such
as \link[stringr:case]{stringr::str_to_sentence}. (N.b. this function must be vectorised).
Any value provided here will be overridden by the
\code{options("tableone.labeller" = my_label_fn)} which allows global setting of
the labeller.}

\item{units}{(optional) a named list of units, following a \verb{c(<colname_1> = "<unit_1>", <colname_2> = "<unit_2>", ...)} format. columns not present in
this list are assumed to have no units. Units may be involved in the
formatting of the summary output.}

\item{override_type}{(optional) a named list of data summary types. The
default type for a column in a data set are calculated using heurisitics
depending on the nature of the data (categorical or continuous), and result
of normality tests. if you want to override this the options are
"subtype_count","median_iqr","mean_sd","skipped" and you
specify this on a column by column bases with a named list (e.g
\code{c("Petal.Width"="mean_sd")}). Overriding the default does not check the
type of data is correct for the summary type and will potentially cause
errors if this is not done correctly.}

\item{override_method}{if you want to override the comparison method for a
particular variable the options are
"chi-sq trend","fisher","t-test","2-sided wilcoxon","2-sided ks","anova","kruskal-wallis","no comparison" and you
specify this on a column by column bases with a named list (e.g
\code{c("Petal.Width"="t-test")})}
}
\value{
a \code{t1_signif} dataframe.
}
\description{
The population comparison is a summary of the co-variates in a data set with
no reference to outcome, but comparing intervention groups. It will report
summary statistics for continuous and counts for categorical data, for each
of the intervention groups, and reports on the significance of the
association in relation to the intervention groups. It gives a clear summary
of whether data is correlated to intervention.
}
\examples{
tmp = iris \%>\% dplyr::group_by(Species) \%>\% as_t1_signif(tidyselect::everything())
tmp = diamonds \%>\% dplyr::group_by(is_colored) \%>\% as_t1_signif(tidyselect::everything())
}
