% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula-utils.R
\name{as_vars}
\alias{as_vars}
\title{Reuse tidy-select syntax outside of a tidy-select function}
\usage{
as_vars(tidyselect, data = NULL)
}
\arguments{
\item{tidyselect}{a tidyselect syntax which will be evaluated in context by looking for a call in the call stack that includes a dataframe as the first argument}

\item{data}{(optional) a specific dataframe with which to evaluate the tidyselect}
}
\value{
a list of symbols resulting from the evaluation of the tidyselect in the context of the current call stack (or a provided data frame)
}
\description{
Reuse tidy-select syntax outside of a tidy-select function
}
