% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bad_test_cols}
\alias{bad_test_cols}
\title{A list of columns for a test case}
\format{
\subsection{\code{bad_test_cols}}{

\describe{Test data}
}
}
\usage{
bad_test_cols
}
\description{
A list of columns for a test case
}
\keyword{datasets}
