% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chk_file.R
\name{chk_file}
\alias{chk_file}
\title{Evaluates that path argument so that if the specified path is NULL a pop window will be displayed to choose file and then the file will be validated for its type.}
\usage{
chk_file(path)
}
\arguments{
\item{path}{(Character) the path that the excel file can be read from or null in case the user wants to have a file choose window displayed.}
}
\value{
the path that the user chose
}
\description{
Evaluates that path argument so that if the specified path is NULL a pop window will be displayed to choose file and then the file will be validated for its type.
}
\author{
Agapios Panos <panosagapios@gmail.com>
}
