% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_population.R
\name{compare_outcomes}
\alias{compare_outcomes}
\title{Compares multiple outcomes against an intervention in a summary table}
\usage{
compare_outcomes(
  df,
  ...,
  label_fn = label_extractor(df),
  units = extract_units(df),
  override_type = list(),
  layout = "compact",
  override_percent_dp = list(),
  override_real_dp = list(),
  p_format = names(.pvalue.defaults),
  font_size = getOption("tableone.font_size", 8),
  font = getOption("tableone.font", "Arial"),
  footer_text = NULL,
  show_binary_value = NULL,
  raw_output = FALSE
)
}
\arguments{
\item{df}{a dataframe of individual observations. If using the \code{tidyselect}
syntax data grouping defines the intervention group and should be present. if the
formula interface is used the first variable in the right hand side of the
formula is used as the intervention, in which case grouping is ignored.}

\item{...}{the outcomes are specified either as a \code{tidyselect} specification,
in which case the grouping of the \code{df} input determines the intervention
and the output is the same as a \code{compare_population()} call with a tidyselect.
Alternatively a set of formulae can be provided that specify the outcomes
on the left hand side, e.g. \verb{outcome1 ~ intervention + cov1, outcome2 ~ intervention + cov1, ...} in this case the \code{intervention} must be the same
for all formulae and used to determine the comparison groups.}

\item{label_fn}{(optional) a function for mapping a co-variate column name to
printable label. This is by default a no-operation and the output table
will contain the dataframe column names as labels. A simple alternative
would be some form of \link[dplyr:case_when]{dplyr::case_when} lookup, or a string function such
as \link[stringr:case]{stringr::str_to_sentence}. (N.b. this function must be vectorised).
Any value provided here will be overridden by the
\code{options("tableone.labeller" = my_label_fn)} which allows global setting of
the labeller.}

\item{units}{(optional) a named list of units, following a \verb{c(<colname_1> = "<unit_1>", <colname_2> = "<unit_2>", ...)} format. columns not present in
this list are assumed to have no units. Units may be involved in the
formatting of the summary output.}

\item{override_type}{(optional) a named list of data summary types. The
default type for a column in a data set are calculated using heurisitics
depending on the nature of the data (categorical or continuous), and result
of normality tests. if you want to override this the options are
"subtype_count","median_iqr","mean_sd","skipped" and you
specify this on a column by column bases with a named list (e.g
\code{c("Petal.Width"="mean_sd")}). Overriding the default does not check the
type of data is correct for the summary type and will potentially cause
errors if this is not done correctly.}

\item{layout}{(optional) various layouts are defined as default. As of this
version of \code{tableone} they are
"relaxed","compact","micro","simple","single","missing". The layouts can be
customised using the options \verb{options("tableone.format_list"=list(...)")},
and this is described in more detail in the vignettes.}

\item{override_percent_dp}{(optional) a named list of overrides for the default
precision of formatting percentages, following a \verb{c(<colname_1> = 2, <colname_2> = 4, ...)} format. columns not present in this list
will use the defaults defined in the layout. See the vignette on customisation.}

\item{override_real_dp}{(optional) a named list of overrides for the default
precision of formatting real values, following a \verb{c(<colname_1> = 2, <colname_2> = 4, ...)} format. columns not present in this list
will use the defaults defined in the layout. See the
\code{utils::vignette("customisation", package="tableone")}.}

\item{p_format}{the format of the p-values: one of
"sampl", "nejm", "jama", "lancet", "aim" but any value
here is overridden by the \verb{option("tableone.pvalue_formatter"=function(...))}}

\item{font_size}{(optional) the font size for the table in points}

\item{font}{(optional) the font family for the table (which will be matched to
closest on your system)}

\item{footer_text}{any text that needs to be added at the end of the table,
setting this to FALSE dsables the whole footer (as does
\code{options("tableone.hide_footer"=TRUE)}).}

\item{show_binary_value}{if set this will filter the display of covariates where the number of possibilities
is exactly 2 to this value.}

\item{raw_output}{return comparison as \code{t1_signif} dataframe rather than formatted table}
}
\value{
a \code{huxtable} formatted table.
}
\description{
The outcome table is a simple summary of a binary or categorical outcome
in a data set compared by intervention groups. The comparison is independent
of any covariates, and is a preliminary output prior to more formal statistical
analysis or model fitting.
}
\details{
It reports summary counts for the outcomes and a measure of significance of
the relationship between outcome and intervention. Interpretation of
significance tests, should include Bonferroni adjustment.
}
