% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableone.R
\name{compare_population}
\alias{compare_population}
\title{Compares the population against an intervention in a summary table}
\usage{
compare_population(
  df,
  ...,
  label_fn = ~.x,
  units = list(),
  override_type = list(),
  layout = "compact",
  override_percent_dp = list(),
  override_real_dp = list(),
  p_format = names(.pvalue.defaults),
  font_size = getOption("tableone.font_size", 8),
  font = getOption("tableone.font", "Arial"),
  footer_text = NULL
)
}
\arguments{
\item{df}{a dataframe of individual observations. If using the \code{tidyselect}
syntax data grouping defines the intervention group and should be present. if the
formula interface is used the first variable in the right hand side of the
formula is used as the intervention, in which case grouping is ignored.}

\item{...}{the columns of variables we wish to summarise. This can be given as
a \code{tidyselect} specification (see \code{utils::vignette("syntax", package = "tidyselect")}),
identifying the columns. Alternatively it can be given as a formula of the nature

\code{outcome ~ intervention + covariate_1 + covariate_2 + ...} .

which may be more convenient if you are going on to do a model fit later. If the
latter format the left hand side is ignored (outcomes are not usual in
this kind of table).}

\item{label_fn}{(optional) a function for mapping a co-variate column name to
printable label. This is by default a no-operation and the output table
will contain the dataframe column names as labels. A simple alternative
would be some form of \link[dplyr:case_when]{dplyr::case_when} lookup, or a string function such
as \link[stringr:case]{stringr::str_to_sentence}. (N.b. this function must be vectorised).
Any value provided here will be overridden by the
\code{options("tableone.labeller" = my_label_fn)} which allows global setting of
the labeller.}

\item{units}{(optional) a named list of units, following a \verb{c(<colname_1> = "<unit_1>", <colname_2> = "<unit_2>", ...)} format. columns not present in
this list are assumed to have no units. Units may be involved in the
formatting of the summary output.}

\item{override_type}{(optional) a named list of data summary types. The
default type for a column in a data set are calculated using heurisitics
depending on the nature of the data (categorical or continuous), and result
of normality tests. if you want to override this the options are
"subtype_count","median_iqr","mean_sd","skipped" and you
specify this on a column by column bases with a named list (e.g
\code{c("Petal.Width"="mean_sd")}). Overriding the default does not check the
type of data is correct for the summary type and will potentially cause
errors if this is not done correctly.}

\item{layout}{(optional) various layouts are defined as default. As of this
version of \code{tableone} they are
"relaxed","compact","simple","single","missing". The layouts can be
customised using the options \verb{options("tableone.format_list"=list(...)")},
and this is described in more detail in the vignettes.}

\item{override_percent_dp}{(optional) a named list of overrides for the default
precision of formatting percentages, following a \verb{c(<colname_1> = 2, <colname_2> = 4, ...)} format. columns not present in this list
will use the defaults defined in the layout. See the vignette on customisation.}

\item{override_real_dp}{(optional) a named list of overrides for the default
precision of formatting real values, following a \verb{c(<colname_1> = 2, <colname_2> = 4, ...)} format. columns not present in this list
will use the defaults defined in the layout. See the
\code{utils::vignette("customisation", package="tableone")}.}

\item{p_format}{the format of the p-values: one of
"sampl", "nejm", "jama", "lancet", "aim" but any value
here is overridden by the \verb{option("tableone.pvalue_formatter"=function(...))}}

\item{font_size}{(optional) the font size for the table in points}

\item{font}{(optional) the font family for the table (which will be matched to
closest on your system)}

\item{footer_text}{any text that needs to be added at the end of the table,
setting this to FALSE dsables the whole footer (as does
\code{options("tableone.hide_footer"=TRUE)}) .}
}
\value{
a \code{huxtable} formatted table.
}
\description{
The population comparison is a summary of the co-variates in a data set with
no reference to outcome, but comparing intervention groups. It will report
summary statistics for continuous and counts for categorical data, for each
of the intervention groups, and reports on the significance of the
association in relation to the intervention groups. It gives a clear summary
of whether data is correlated to intervention.
}
\examples{
# the heuristics detect that Petals in the iris data set are not normally
# distributed and hence report median and IQR:
iris \%>\% dplyr::group_by(Species) \%>\% compare_population(tidyselect::everything())

# Missing data
old = options("tableone.show_pvalue_method"=TRUE)
missing_diamonds \%>\% dplyr::group_by(is_colored) \%>\% compare_population(-color, layout="relaxed")

tmp = missing_diamonds \%>\% explicit_na() \%>\% dplyr::group_by(is_colored)
tmp \%>\% compare_population(-color,
    footer_text = c(
      "IQR: Interquartile range; CI: Confidence interval",
      "Line two")
    )

options(old)
}
