% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableone.R
\name{cut_integer}
\alias{cut_integer}
\title{Cut and label an integer valued quantity}
\usage{
cut_integer(
  x,
  cut_points,
  glue = "{label}",
  lower_limit = -Inf,
  upper_limit = Inf,
  ...
)
}
\arguments{
\item{x}{a vector of integer valued numbers, e.g. ages, counts}

\item{cut_points}{a vector of integer valued cut points which define the lower, inclusive boundary of each group}

\item{glue}{a glue spec that may be used to generate a label. It can use {low}, {high}, {next_low}, or {label} as values.}

\item{lower_limit}{the minimum value we should include (this is inclusive for the bottom category) (default -Inf)}

\item{upper_limit}{the maximum value we should include (this is also inclusive for the top category) (default Inf)}

\item{...}{not used}
}
\value{
an ordered factor of the integer
}
\description{
Deals with some annoying issues classifying integer data sets, such as ages, into groups. where you want to
specify just the change over points as integers and clearly label the resulting ordered factor.
}
\examples{
cut_integer(rbinom(20,20,0.5), c(5,10,15))
cut_integer(floor(runif(100,-10,10)), cut_points = c(2,3,4,6), lower_limit=0, upper_limit=10)
cut_integer(1:10, cut_points = c(1,3,9))
}
