% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{default.format}
\alias{default.format}
\title{Default table layout functions}
\format{
\subsection{\code{default.format}}{

A names list of lists:
\describe{
\item{level one}{The name of the table layout}
\item{level two}{The name of the summary type required. one of \code{subtype_count},
\code{median_iqr},\code{mean_sd},\code{skipped}}
\item{level three}{a named list of \code{column}=\verb{glue specification} pairs. The
\code{column} (itself a glue spec) might reference \code{N_total}, \code{N_present} or \code{.unit} but
typically will be a fixed string- it defines the name of the table column
to generate. The \verb{glue specification} defines the layout of that column,
and can use summary statistics as below}
\item{subtype_count}{can use \code{level}, \code{prob.0.5}, \code{prob.0.025},
\code{prob.0.975}, \code{unit}, \code{n}, \code{N}. \code{n} is subgroup count, \code{N} is data count.}
\item{median_iqr}{can use \code{q.0.5}, \code{q.0.25}, ..., \code{unit}, \code{n}, \code{N} - \code{n} excludes
missing, \code{N} does not.}
\item{mean_sd}{can use \code{mean}, \code{sd}, \code{unit}, \code{n}, \code{N} - \code{n} excludes
missing, \code{N} does not.}
\item{skipped}{can use \code{n}, \code{N} - \code{n} excludes
missing, \code{N} does not.}
}
}
}
\usage{
default.format
}
\description{
Customisation of output can use one of these entries as a starting point.
A custom layout should look like one of the entries in level 2 of this
nested list, containing 4 named entries, one for each type of table summary.
}
\keyword{datasets}
