% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe_data.R
\name{describe_data}
\alias{describe_data}
\title{Describe the data types and consistence}
\usage{
describe_data(
  df,
  ...,
  label_fn = label_extractor(df),
  units = extract_units(df),
  layout = "single",
  font_size = getOption("tableone.font_size", 8),
  font = getOption("tableone.font", "Arial"),
  footer_text = NULL,
  raw_output = FALSE
)
}
\arguments{
\item{df}{a dataframe of individual observations. Grouping, if present, is ignored.
(n.b. if you wanted to construct multiple summary tables a \code{\link[dplyr:group_map]{dplyr::group_map()}} call
could be used)}

\item{...}{the columns of variables we wish to summarise. This can be given as
a \code{tidyselect} specification (see \code{utils::vignette("syntax", package = "tidyselect")}),
identifying the columns. Alternatively it can be given as a formula of the nature

\code{outcome ~ intervention + covariate_1 + covariate_2 + ...} .

which may be more convenient if you are going on to do a model fit. If the
latter format the left hand side is ignored (outcomes are not usual in
this kind of table).}

\item{label_fn}{(optional) a function for mapping a co-variate column name to
printable label. This is by default a no-operation and the output table
will contain the dataframe column names as labels. A simple alternative
would be some form of \link[dplyr:case_when]{dplyr::case_when} lookup, or a string function such
as \link[stringr:case]{stringr::str_to_sentence}. (N.b. this function must be vectorised).
Any value provided here will be overridden by the
\code{options("tableone.labeller" = my_label_fn)} which allows global setting of
the labeller.}

\item{units}{(optional) a named list of units, following a \verb{c(<colname_1> = "<unit_1>", <colname_2> = "<unit_2>", ...)} format. columns not present in
this list are assumed to have no units. Units may be involved in the
formatting of the summary output.}

\item{layout}{(optional) various layouts are defined as default. As of this
version of \code{tableone} they are
"relaxed","compact","micro","simple","single","missing". The layouts can be
customised using the options \verb{options("tableone.format_list"=list(...)")},
and this is described in more detail in the vignettes.}

\item{font_size}{(optional) the font size for the table in points}

\item{font}{(optional) the font family for the table (which will be matched to
closest on your system)}

\item{footer_text}{any text that needs to be added at the end of the table,
setting this to FALSE dsables the whole footer (as does
\code{options("tableone.hide_footer"=TRUE)}).}

\item{raw_output}{return comparison as \code{t1_signif} dataframe rather than formatted table}
}
\value{
a \code{huxtable} formatted table.
}
\description{
The population description is a simple summary of the co-variates in a data set
with no reference to outcome, and not comparing intervention (although it might
contain intervention rates.) It will report summary statistics for continuous
and counts for categorical data,
}
\examples{
# Overriding the heuristics is possible:
iris \%>\% describe_data(tidyselect::everything())

diamonds \%>\% dplyr::group_by(is_colored) \%>\% describe_data(tidyselect::everything())

}
