% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{diamonds}
\alias{diamonds}
\title{A copy of the diamonds dataset}
\format{
\subsection{\code{diamonds}}{

\describe{Test data}
}
}
\usage{
diamonds
}
\description{
with a binary class is_coloured based on the color column
}
\keyword{datasets}
