% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableone.R
\name{explicit_na}
\alias{explicit_na}
\title{Make NA values in factor columns explicit}
\usage{
explicit_na(df, na_level = "<missing>")
}
\arguments{
\item{df}{the data frame}

\item{na_level}{a label for NA valued factors}
}
\value{
the dataframe with all factor columns containing explicit na values
}
\description{
Converts NA values in any factors in the dataframe into a new level -
This is a thin wrapper for \code{\link[forcats:fct_explicit_na]{forcats::fct_explicit_na()}}
}
\examples{
# before
missing_diamonds \%>\% dplyr::group_by(cut) \%>\% dplyr::count()
# after
missing_diamonds \%>\% explicit_na() \%>\% dplyr::group_by(cut) \%>\% dplyr::count()
}
