% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-utils.R
\name{extract_units}
\alias{extract_units}
\title{Extracts units set as dataframe column attributes}
\usage{
extract_units(df)
}
\arguments{
\item{df}{the data frame from \code{set_units()}}
}
\value{
a named list of column / unit pairs.
}
\description{
Extracts units set as dataframe column attributes
}
\examples{
iris = iris \%>\% set_units(-Species, units="mm")
iris \%>\% extract_units()
}
