% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ExtractSmd.R
\name{ExtractSmd}
\alias{ExtractSmd}
\title{Extract standardized mean differences from a (svy)TableOne object}
\usage{
ExtractSmd(x)
}
\arguments{
\item{x}{A stratified (svy)TableOne object containing standardized mean differences.}
}
\value{
A matrix containing the average standardized mean differences as well as the all possible pairwise standardized mean differences. Variables are ordered in the same order as the printed table.
}
\description{
Extracts standardized mean differences data as a matrix from a (svy)TableOne object
}
\examples{
##
}
\author{
Kazuki Yoshida
}

