% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TABLEONE.r
\name{find_cell}
\alias{find_cell}
\title{Calculate numbers and statistics for table one}
\usage{
find_cell(g.obj, row, col, name = "core-fg")
}
\arguments{
\item{g.obj}{A tableGrob object.}

\item{row}{row id.}

\item{col}{column id.}

\item{name}{default option}
}
\value{
indexes of cells matching criteria.
}
\description{
Calculate numbers and statistics for table one
}
