% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableone.R
\name{get_footer_text}
\alias{get_footer_text}
\title{Get footer text if available}
\usage{
get_footer_text(df_output)
}
\arguments{
\item{df_output}{a data frame that is the output of a \code{tableone} function}
}
\value{
the footnotes if they exist as a list (NULL otherwise)
}
\description{
The functions in \code{tableone} will record the methods used for reporting in a
scientific paper. This is both for normality assumption tests and for
significance tests.
}
\examples{
iris \%>\% describe_population(tidyselect::everything()) \%>\% get_footer_text()
iris \%>\% dplyr::group_by(Species) \%>\%
  compare_population(tidyselect::everything()) \%>\% get_footer_text()
}
