% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_excel_data.R
\name{import_excel_data}
\alias{import_excel_data}
\title{Imports the data from the excel file}
\usage{
import_excel_data(path = NULL, sheet = NULL, import.col.names,
  output.var.names, group.col.name, control.value, treatment.value)
}
\arguments{
\item{path}{(Character) the path that the excel file can be read from. By default it is set to NULL so that a pop up window will ask for the path.}

\item{sheet}{(Character) (default: NULL) the sheet inside excel file that the data are stored. By default it gets the first one.}

\item{import.col.names}{(Character) the column names as specifed in the excel.}

\item{output.var.names}{(Character) the variable names as they will be displayed in the output.}

\item{group.col.name}{(Character) the column name that specifies the group (control or treatment)}

\item{control.value}{(Character) the value that specifies the control group. (e.g. 0 or 'control')}

\item{treatment.value}{(Character) the value that specifies the treatment group (e.g. 1 or 'treatment')}
}
\value{
data.frame with the data.
}
\description{
Imports the data from the excel file
}
\author{
Agapios Panos <panosagapios@gmail.com>
}
