% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{index}
\alias{index}
\title{Toy example for descriptive data}
\format{
A data frame with 235 rows and 6 variables:
#' \describe{
\item{RIN}{RIN}
\item{Gender}{Gender}
\item{Age}{Age}
\item{Packyear}{Packyear}
\item{Smoking}{Smoking Status}
\item{Cancer}{Cancer Status}
}
}
\usage{
index
}
\description{
Toy example for descriptive data
}
\keyword{datasets}
