% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-utils.R
\name{label_extractor}
\alias{label_extractor}
\title{Extract labels from a dataframe column attributes}
\usage{
label_extractor(df, ..., attribute = "label")
}
\arguments{
\item{df}{a dataframe containing some labels}

\item{...}{additional string manipulation functions to apply e.g. \code{tolower}}

\item{attribute}{the name of the label containing attribute (defaults to \code{"label"})}
}
\value{
a labelling function. This is specific to the dataframe provided in \code{df}
}
\description{
Retrieve column labels are embedded as an attribute of each column.
}
\examples{
iris = set_labels(iris, c(
    "Sepal Length", "Sepal Width",
    "Petal Length", "Petal Width",  "Species"
 ))
fn = label_extractor(iris,tolower)
fn(colnames(iris))
}
