% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{missing_diamonds}
\alias{missing_diamonds}
\title{A copy of the diamonds dataset}
\format{
\subsection{\code{missing_diamonds}}{

\describe{Test data}
}
}
\usage{
missing_diamonds
}
\description{
with 10\% of entries replaced by NA
and a binary class is_coloured based on the color column
}
\keyword{datasets}
