% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mnar_two_class_1000}
\alias{mnar_two_class_1000}
\title{Missing not at random 2 class 1000 items}
\format{
\subsection{\code{mnar_two_class_1000}}{

\describe{Test data}
}
}
\usage{
mnar_two_class_1000
}
\description{
A random data test dataset with 2 classes (groupings column)
one of which has 10\% missing data and the other has 20\%
}
\keyword{datasets}
