% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_csv_fileInput.R
\name{mod_csv_fileInput}
\alias{mod_csv_fileInput}
\alias{mod_csv_file}
\title{mod_csv_fileInput and mod_csv_file}
\usage{
mod_csv_fileInput(id, label = "CSV file")

mod_csv_file(input, output, session, stringsAsFactors = TRUE)
}
\arguments{
\item{id}{shiny id}

\item{label}{fileInput label}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module that imports csv file
}
