% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_export_word_table.R
\name{mod_export_word_table_ui}
\alias{mod_export_word_table_ui}
\alias{mod_export_word_table_server}
\title{mod_export_word_table_ui and mod_export_word_table_server}
\usage{
mod_export_word_table_ui(id)

mod_export_word_table_server(input, output, session, data)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
