% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_inputFile.R
\name{mod_inputFile_ui}
\alias{mod_inputFile_ui}
\alias{mod_inputFile_server}
\title{mod_inputFile_ui and mod_inputFile_server}
\usage{
mod_inputFile_ui(id, label = "file")

mod_inputFile_server(input, output, session, stringsAsFactors)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
