% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_TableOne.R
\name{mod_tableOneInput}
\alias{mod_tableOneInput}
\alias{mod_TableOne_server}
\title{mod_TableOne_ui and mod_TableOne_server}
\usage{
mod_tableOneInput(id, choices, multiple = TRUE)

mod_TableOne_server(input, output, session, df)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
