% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{multi_class_negative}
\alias{multi_class_negative}
\title{A multi-class dataset with equal random samples in each class}
\format{
\subsection{\code{multi_class_negative}}{

\describe{Test data}
}
}
\usage{
multi_class_negative
}
\description{
A multi-class dataset with equal random samples in each class
}
\keyword{datasets}
