% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numericSummary.R
\name{numericSummary}
\alias{numericSummary}
\title{numericSummary provides the table text for summarizing a numeric vector}
\usage{
numericSummary(vec2Summarize, func1 = mean, func2 = sd, sigInts = 2)
}
\arguments{
\item{vec2Summarize}{- the numeric vector to summarize}

\item{func1}{Defaults to mean; any function working on a numeric vector can be used}

\item{func2}{Defaults to sd; any function working on a numeric vector can be used}

\item{sigInts}{Defaults to 2; the number of significant integers to print in the returned character}
}
\description{
This function allows text summarization of a numeric vector. The summary functions can be specified as can the number of significant digits returned
}
\examples{
numericSummary(1:10)
}

