% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{one_class_test_1000}
\alias{one_class_test_1000}
\title{A single-class dataset with 1000 items of random data}
\format{
\subsection{\code{one_class_test_1000}}{

\describe{Test data}
}
}
\usage{
one_class_test_1000
}
\description{
columns contain a set of random data of different types e.g.
uniform continuous, normal, binomial, multinomial.
}
\keyword{datasets}
