% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/print.svyCatTable.R
\name{print.svyCatTable}
\alias{print.svyCatTable}
\title{Format and print the \code{svyCatTable} class objects}
\usage{
\method{print}{svyCatTable}(x, digits = 1, pDigits = 3, quote = FALSE,
  missing = FALSE, explain = TRUE, printToggle = TRUE, noSpaces = FALSE,
  format = c("fp", "f", "p", "pf")[1], showAllLevels = FALSE,
  cramVars = NULL, test = TRUE, exact = NULL, CrossTable = FALSE, ...)
}
\arguments{
\item{x}{The result of a call to the \code{\link{svyCreateCatTable}} function.}

\item{digits}{Number of digits to print in the table.}

\item{pDigits}{Number of digits to print for p-values.}

\item{quote}{Whether to show everything in quotes. The default is FALSE. If TRUE, everything including the row and column names are quoted so that you can copy it to Excel easily.}

\item{missing}{Whether to show missing data information (not implemented yet, placeholder)}

\item{explain}{Whether to add explanation to the variable names, i.e., (\%) is added to the variable names when percentage is shown.}

\item{printToggle}{Whether to print the output. If FLASE, no output is created, and a matrix is invisibly returned.}

\item{noSpaces}{Whether to remove spaces added for alignment. Use this option if you prefer to align numbers yourself in other software.}

\item{format}{The default is "fp" frequency (percentage). You can also choose from "f" frequency only, "p" percentage only, and "pf" percentage (frequency).}

\item{showAllLevels}{Whether to show all levels. FALSE by default, i.e., for 2-level categorical variables, only the higher level is shown to avoid redundant information.}

\item{cramVars}{A character vector to specify the two-level categorical variables, for which both levels should be shown in one row.}

\item{test}{Whether to show the p-values. TRUE by default. If FALSE, only the numerical summaries are shown.}

\item{exact}{This option is not available for tables from weighted data.}

\item{CrossTable}{Whether to show the cross table objects held internally using gmodels::CrossTable function. This will give an output similar to the PROC FREQ in SAS.}

\item{...}{For compatibility with generic. Ignored.}
}
\value{
It is mainly for printing the result. But this function does return a matrix containing what you see in the output invisibly. You can assign it to an object to save it.
}
\description{
This is the \code{print} method for the \code{svyCatTable} class objects created by \code{\link{svyCreateCatTable}} function.
}
\examples{
## See the examples for svyCreateTableOne()
}
\author{
Kazuki Yoshida
}
\seealso{
\code{\link{svyCreateCatTable}},  \code{\link{print.svyCatTable}},  \code{\link{summary.svyCatTable}},
\code{\link{svyCreateContTable}}, \code{\link{print.svyContTable}}, \code{\link{summary.svyContTable}},
\code{\link{svyCreateTableOne}},  \code{\link{print.TableOne}},     \code{\link{summary.TableOne}}
}

