% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/print.svyContTable.R
\name{print.svyContTable}
\alias{print.svyContTable}
\title{Format and print the \code{svyContTable} class objects}
\usage{
\method{print}{svyContTable}(x, digits = 2, pDigits = 3, quote = FALSE,
  missing = FALSE, explain = TRUE, printToggle = TRUE, noSpaces = FALSE,
  nonnormal = NULL, minMax = FALSE, insertLevel = FALSE, test = TRUE,
  ...)
}
\arguments{
\item{x}{The result of a call to the \code{\link{svyCreateContTable}} function.}

\item{digits}{Number of digits to print in the table.}

\item{pDigits}{Number of digits to print for p-values.}

\item{quote}{Whether to show everything in quotes. The default is FALSE. If TRUE, everything including the row and column names are quoted so that you can copy it to Excel easily.}

\item{missing}{Whether to show missing data information (not implemented yet, placeholder)}

\item{explain}{Whether to add explanation to the variable names, i.e., (mean (sd) or median [IQR]) is added to the variable names.}

\item{printToggle}{Whether to print the output. If FLASE, no output is created, and a matrix is invisibly returned.}

\item{noSpaces}{Whether to remove spaces added for alignment. Use this option if you prefer to align numbers yourself in other software.}

\item{nonnormal}{A character vector to specify the variables for which the p-values should be those of nonparametric tests. By default all p-values are from normal assumption-based tests (oneway.test).}

\item{minMax}{Whether to use [min,max] instead of [p25,p75] for nonnormal variables. The default is FALSE.}

\item{insertLevel}{Whether to add an empty level column to the left of strata.}

\item{test}{Whether to show the p-values. TRUE by default. If FALSE, only the numerical summaries are shown.}

\item{...}{For compatibility with generic. Ignored.}
}
\value{
It is mainly for printing the result. But this function does return a matrix containing what you see in the output invisibly. You can assign it to an object to save it.
}
\description{
This is the \code{print} method for the \code{svyContTable} class objects created by \code{\link{CreateContTable}} function.
}
\examples{
## See the examples for svyCreateTableOne()
}
\author{
Kazuki Yoshida
}
\seealso{
\code{\link{svyCreateCatTable}},  \code{\link{print.svyCatTable}},  \code{\link{summary.svyCatTable}},
\code{\link{svyCreateContTable}}, \code{\link{print.svyContTable}}, \code{\link{summary.svyContTable}},
\code{\link{svyCreateTableOne}},  \code{\link{print.TableOne}},     \code{\link{summary.TableOne}}
}

