% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/returnPVal.R
\name{returnPVal}
\alias{returnPVal}
\title{returnPVal - a helper function to return table1 p-values
 
This function returns the p-values for numeric and factor based comparisons between the groups}
\usage{
returnPVal(eSet, groupChar, char, numericTest = aov,
  factorTest = fisher.test)
}
\arguments{
\item{eSet}{- the eSet or data.frame sent to table1}

\item{groupChar}{- a factor column present in eSet by which to group samples and test char}

\item{char}{- which characteristic is summarized from eSet}

\item{numericTest}{Defaults to aov - changing this doesn't work yet}

\item{factorTest}{Defaults to fisher.test - changing this doesn't work yet}
}
\description{
returnPVal - a helper function to return table1 p-values
 
This function returns the p-values for numeric and factor based comparisons between the groups
}
\examples{
returnNumSlots(eSet, "CancerStatus", "CancerPositive", "Age")
}

