% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/returnSlots.R
\name{returnSlots}
\alias{returnSlots}
\title{returnSlots - a helper function to return slots to place in the table1
 
This function returns the slots generated by returnFactorSlots and returnNumSlots}
\usage{
returnSlots(eSet, groupChar, group, char, func = NULL)
}
\arguments{
\item{eSet}{- the eSet or data.frame sent to table1}

\item{groupChar}{- a factor column present in eSet by which to group samples and summarize char}

\item{group}{- which factor level from groupChar is summarized}

\item{char}{- which characteristic is summarized from eSet}

\item{func}{- a user defined function that will also return a character summary - not really implemented yet}
}
\description{
returnSlots - a helper function to return slots to place in the table1
 
This function returns the slots generated by returnFactorSlots and returnNumSlots
}
\examples{
returnNumSlots(eSet, "CancerStatus", "CancerPositive", "Age")
}

