% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/returnSlotsUnique.R
\name{returnSlotsUnique}
\alias{returnSlotsUnique}
\title{returnSlotsUnique - a helper function to help generate slots for table based on single patients/samples instead of groups (i.e. that in Campbell 2012)
 
This function returns a character representing the results of the summarization; meant for use when generating tables for patients instead of groups}
\usage{
returnSlotsUnique(eSet, groupChar, group, char)
}
\arguments{
\item{eSet}{- the eSet or data.frame sent to table1}

\item{groupChar}{- a factor column present in eSet by which to group samples and summarize char}

\item{group}{- which factor level from groupChar is summarized}

\item{char}{- which characteristic is summarized from eSet}
}
\description{
returnSlotsUnique - a helper function to help generate slots for table based on single patients/samples instead of groups (i.e. that in Campbell 2012)
 
This function returns a character representing the results of the summarization; meant for use when generating tables for patients instead of groups
}
\examples{
returnSlotsUnique(eSet, "Patient.ID", "Patient7031", "SmokingStatus")
}

