% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roundN.R
\name{roundN}
\alias{roundN}
\title{Round a Number}
\usage{
roundN(n)
}
\arguments{
\item{p}{a numeric vector of p-values to round}
}
\value{
a numeric vector of rounded p-values
}
\description{
Round a Number
}
\examples{
library(tableOne)
p <- c(runif(10), pmin(rexp(10,20),0.5))
cbind(p, roundP(p))
}
