% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roundP.R
\name{roundP}
\alias{roundP}
\title{Round a P-Value}
\usage{
roundP(p)
}
\arguments{
\item{p}{a numeric vector of p-values to round}
}
\value{
a numeric vector of rounded p-values
}
\description{
Round a P-Value
}
\examples{
library(tableOne)
p <- c(runif(10), pmin(rexp(10,20),0.5))
cbind(p, roundP(p))
}
