% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-utils.R
\name{set_units}
\alias{set_units}
\title{Title}
\usage{
set_units(df, ..., units)
}
\arguments{
\item{df}{a dataframe}

\item{...}{a tidyselect specification or a formula}

\item{units}{a list of unit as strings which must be either 1 or the same length as
the columns matched by the tidyselect.}
}
\value{
the dataframe with the \code{unit} attribute updated
}
\description{
Title
}
\examples{
iris = iris \%>\% set_units(-Species, units="mm")
iris \%>\% extract_units()
}
