% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/summary.svyCatTable.R
\name{summary.svyCatTable}
\alias{summary.svyCatTable}
\title{Shows all results in a \code{svyCatTable} class object}
\usage{
\method{summary}{svyCatTable}(object, digits = 1, ...)
}
\arguments{
\item{object}{An object that has the \code{svyCatTable} class to be shown.}

\item{digits}{Number of digits to print.}

\item{...}{For compatibility with generic. Ignored.}
}
\value{
It will print the results.
}
\description{
This method shows all the data a \code{svyCatTable} class object has. This includes the (optionally stratified) part with summary statistics and , if available, p-values from the approximation method test (\code{\link{chisq.test}} by default) and exact method test (\code{\link{fisher.test}} by default).
}
\examples{
## See the examples for svyCreateTableOne()
}
\author{
Kazuki Yoshida
}
\seealso{
\code{\link{svyCreateCatTable}},  \code{\link{print.svyCatTable}},  \code{\link{summary.svyCatTable}},
\code{\link{svyCreateContTable}}, \code{\link{print.svyContTable}}, \code{\link{summary.svyContTable}},
\code{\link{svyCreateTableOne}},  \code{\link{print.TableOne}},     \code{\link{summary.TableOne}}
}

