% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/summary.svyContTable.R
\name{summary.svyContTable}
\alias{summary.svyContTable}
\title{Shows all results in a \code{svyContTable} class object}
\usage{
\method{summary}{svyContTable}(object, digits = 2, ...)
}
\arguments{
\item{object}{An object that has the \code{svyContTable} class to be shown.}

\item{digits}{Number of digits to print.}

\item{...}{For compatibility with generic. Ignored.}
}
\value{
It will print the results.
}
\description{
This method shows all the data a \code{svyContTable} class object has. This includes the (optionally stratified) part with summary statistics and , if available, p-values from the normal assupmtion-based test (\code{\link{oneway.test}} by default) and nonparametric test (\code{\link{kruskal.test}} by default).
}
\examples{
## See the examples for svyCreateTableOne()
}
\author{
Kazuki Yoshida
}
\seealso{
\code{\link{svyCreateContTable}}, \code{\link{print.svyContTable}}, \code{\link{summary.svyContTable}},
\code{\link{svyCreateCatTable}},  \code{\link{print.svyCatTable}},  \code{\link{summary.svyCatTable}},
\code{\link{svyCreateTableOne}},  \code{\link{print.TableOne}},  \code{\link{summary.TableOne}}
}

