% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TABLEONE.r
\name{tableone.export}
\alias{tableone.export}
\title{Write table one to excel file}
\usage{
tableone.export(table.obj, filename = "Table1")
}
\arguments{
\item{table.obj}{A output from the tableone() function.}

\item{filename}{A character indicating the output name of the files. Default is table1.}
}
\value{
A xlsx file of table one.
}
\description{
Write table one to excel file
}
