% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TABLEONE.r
\name{tableone.plot}
\alias{tableone.plot}
\title{Plot table one}
\usage{
tableone.plot(table.obj, p.threshold = 0.05)
}
\arguments{
\item{table.obj}{A output from the tableone() function.}

\item{p.threshold}{A numeric indicating significant value cutoff for bolding p-value.}
}
\value{
A plot of table one.
}
\description{
Plot table one
}
