% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableOne_oneColumn.R
\name{tableOne_oneColumn}
\alias{tableOne_oneColumn}
\title{Build Table One}
\usage{
tableOne_oneColumn(data, pretty.labels)
}
\arguments{
\item{data}{a data frame to summarize}

\item{pretty.labels}{a named character vector where the names correspond to all
column names of data except for the strata variable. Values correspond to the
cleaned version of the variable names to be displayed in the summary table.}
}
\value{
a tibble with summary statistics (mean/sd or n/%)
}
\description{
Build Table One
}
\examples{
library(tableOne)
data(iris)
data <- iris
tableOne(
    iris,
    pretty.labels = c(
        Species = "Species",
        Sepal.Length = "Sepal Length (cm)",
        Sepal.Width = "Sepal Width (cm)",
        Petal.Length = "Petal Length (cm)"),
        Petal.Width = "Petal Width (cm)"))
}
