% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableRowNum.R
\name{tableRowNum}
\alias{tableRowNum}
\title{tableRowNum - a helper function to determine the number of rows needed in table1
 
This function returns the number of rows that will be needed for table 1 based on the variables that will be summarized}
\usage{
tableRowNum(eSet, characteristics)
}
\arguments{
\item{eSet}{- the eSet or data.frame sent to table1}

\item{characteristics}{- which characteristics are summarized from eSet}
}
\description{
tableRowNum - a helper function to determine the number of rows needed in table1
 
This function returns the number of rows that will be needed for table 1 based on the variables that will be summarized
}
\examples{
tableRowNum(eSet, c("CancerStatus", "SmokingStatus", "PackYears"))
}

