% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{two_class_test}
\alias{two_class_test}
\title{A two-class dataset with random data}
\format{
\subsection{\code{one_class_test_100}}{

\describe{Test data}
}
}
\usage{
two_class_test
}
\description{
columns contain a set of random data of different types e.g.
uniform continuous, normal, binomial, multinomial. in grouping 1
there is 100 items in grouping 2 there are 1000 items
}
\keyword{datasets}
