% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableone.R
\name{compare_variables}
\alias{compare_variables}
\title{Get summary comparisons and statistics between variables as text.}
\usage{
compare_variables(df, ..., override_type = list())
}
\arguments{
\item{df}{a dataframe of individual observations. If using the \code{tidyselect}
syntax data grouping defines the intervention group and should be present. if the
formula interface is used the first variable in the right hand side of the
formula is used as the intervention, in which case grouping is ignored.}

\item{...}{the outcomes are specified either as a \code{tidyselect} specification, in
which case the grouping of the \code{df} input determines the intervention and
the output is the same a \code{compare_population()} call with a tidyselect.
Alternatively a set of formulae can be provided that specify the outcomes
on the left hand side, e.g. \verb{outcome1 ~ intervention + cov1, outcome2 ~ intervention + cov1, ...}
in this case the \code{intervention} must be the same for all formulae and used
to determine the comparison groups.}

\item{override_type}{(optional) a named list of data summary types. The
default type for a column in a data set are calculated using heurisitics
depending on the nature of the data (categorical or continuous), and result
of normality tests. if you want to override this the options are
"subtype_count","median_iqr","mean_sd","skipped" and you
specify this on a column by column bases with a named list (e.g
\code{c("Petal.Width"="mean_sd")}). Overriding the default does not check the
type of data is correct for the summary type and will potentially cause
errors if this is not done correctly.}
}
\value{
a \code{huxtable} formatted table.
}
\description{
Get summary comparisons and statistics between variables as text.
}
