% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.CatTable.R
\name{summary.CatTable}
\alias{summary.CatTable}
\title{Shows all results in a \code{CatTable} class object}
\usage{
\method{summary}{CatTable}(object, digits = 1, ...)
}
\arguments{
\item{object}{An object that has the \code{CatTable} class to be shown.}

\item{digits}{Number of digits to print.}

\item{...}{For compatibility with generic. Ignored.}
}
\value{
None. Results are printed.
}
\description{
Shows all data a \code{CatTable} class object has. This includes the (optionally stratified) part with summary statistics and, if available, p-values from the approximation method test (\code{chisq.test} by default) and exact method test (\code{fisher.test} by default) and standardized mean differences of all possible pairwise contrasts.
}
\examples{

## See examples for CreateTableOne

}
\seealso{
\code{\link{CreateTableOne}}, \code{\link{CreateCatTable}}, \code{\link{print.CatTable}},
}
\author{
Kazuki Yoshida
}
