% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.ContTable.R
\name{summary.ContTable}
\alias{summary.ContTable}
\title{Shows all results in a \code{ContTable} class object}
\usage{
\method{summary}{ContTable}(object, digits = 2, ...)
}
\arguments{
\item{object}{An object that has the \code{ContTable} class to be shown.}

\item{digits}{Number of digits to print.}

\item{...}{For compatibility with generic. Ignored.}
}
\value{
None. Results are printed.
}
\description{
Shows all data a \code{ContTable} class object has. This includes the (optionally stratified) part with summary statistics and, if available, p-values from the normal assupmtion-based test (\code{oneway.test} by default) and nonparametric test (\code{kruskal.test} by default) and standardized mean differences of all possible pairwise contrasts.
}
\examples{

## See examples for CreateTableOne

}
\seealso{
\code{\link{CreateTableOne}}, \code{\link{CreateContTable}}, \code{\link{print.ContTable}}
}
\author{
Kazuki Yoshida
}
