% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.svyCatTable.R
\name{summary.svyCatTable}
\alias{summary.svyCatTable}
\title{Shows all results in a \code{svyCatTable} class object}
\usage{
\method{summary}{svyCatTable}(object, digits = 1, ...)
}
\arguments{
\item{object}{An object that has the \code{svyCatTable} class to be shown.}

\item{digits}{Number of digits to print.}

\item{...}{For compatibility with generic. Ignored.}
}
\value{
None. Results are printed.
}
\description{
Shows all data a \code{svyCatTable} class object has. This includes the (optionally stratified) part with summary statistics and, if available, p-values from the approximation method test (\code{chisq.test} by default) and exact method test (\code{fisher.test} by default) and standardized mean differences of all possible pairwise contrasts.
}
\examples{

## See the examples for svyCreateTableOne()

}
\seealso{
\code{\link{svyCreateTableOne}}, \code{\link{svyCreateCatTable}},  \code{\link{print.svyCatTable}}
}
\author{
Kazuki Yoshida
}
