% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.svyContTable.R
\name{summary.svyContTable}
\alias{summary.svyContTable}
\title{Shows all results in a \code{svyContTable} class object}
\usage{
\method{summary}{svyContTable}(object, digits = 2, ...)
}
\arguments{
\item{object}{An object that has the \code{svyContTable} class to be shown.}

\item{digits}{Number of digits to print.}

\item{...}{For compatibility with generic. Ignored.}
}
\value{
None. Results are printed.
}
\description{
Shows all data a \code{svyContTable} class object has. This includes the (optionally stratified) part with summary statistics and, if available, p-values from the normal assupmtion-based test (\code{regTermTest} with \code{svyglm} by default) and nonparametric test (\code{svyranktest} by default) and standardized mean differences of all possible pairwise contrasts.
}
\examples{

## See the examples for svyCreateTableOne()

}
\seealso{
\code{\link{svyCreateTableOne}}, \code{\link{svyCreateContTable}}, \code{\link{print.svyContTable}}
}
\author{
Kazuki Yoshida
}
