% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableone.R
\name{tableone}
\alias{tableone}
\title{Exports the table 1 in a word file}
\usage{
tableone(columns, var.names, percentages, group.col.name, control.value,
  treatment.value, path = NULL, export.path = NULL, sheet = NULL,
  tableone.col.names = NULL, export.filename = NULL)
}
\arguments{
\item{columns}{(Character) the column names as specifed in the excel.}

\item{var.names}{(Character) the variable names as they will be displayed in the output.}

\item{percentages}{(Boolean) a vector of boolean values that specifies whether that exported variable value will be displayed in a \% format.}

\item{group.col.name}{(Character) the column name that specifies the group (control or treatment).}

\item{control.value}{(Character) the value that specifies the control group. (e.g. 0 or 'control').}

\item{treatment.value}{(Character) the value that specifies the treatment group (e.g. 1 or 'treatment').}

\item{path}{(Character) (optional) the path that the excel file can be read from. By default it is set to NULL so that a pop up window will ask for the path.}

\item{export.path}{(Character) (optional) the path that the Word Document will be exported to.}

\item{sheet}{(Character) (optional) (default: NULL) the sheet inside excel file that the data are stored. By default it gets the first one.}

\item{tableone.col.names}{(Character) (optional) a vector for the column names of the exported table. Default are: c(' ', 'Treatment Group', 'Control Group', '95\% CI', 'Z-value', 'p-value')}

\item{export.filename}{(Character) (optional) the name of the file that will be exported. Do not include the .docx extension. (default filename is TableOne.docx)}
}
\description{
Exports the table 1 in a word file
}
\author{
Agapios Panos <panosagapios@gmail.com>
}
