% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableOne.R
\name{tableOne}
\alias{tableOne}
\title{Build Table One}
\usage{
tableOne(data, strata.variable, pretty.labels, includeMarginal)
}
\arguments{
\item{data}{a data frame to summarize}

\item{strata.variable}{a single character naming the column in the data frame which contains a character/factor of the strata}

\item{pretty.labels}{a named character vector where the names correspond to all
column names of data except for the strata variable. Values correspond to the
cleaned version of the variable names to be displayed in the summary table.}
}
\value{
a tibble with summary statistics (mean/sd or n/%)
}
\description{
Build Table One
}
\examples{
library(tableOne)
data(iris)
data <- iris
data$species2 <-  data$Species
tableOne(
    iris,
    strata = "Species",
    pretty.labels = c(
        Sepal.Length = "Sepal Length (cm)",
        Sepal.Width = "Sepal Width (cm)",
        Petal.Length = "Petal Length (cm)"),
        Petal.Width = "Petal Width (cm)"))
}
