% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TABLEONE.r
\name{tableone}
\alias{tableone}
\title{Calculate numbers and statistics for table one}
\usage{
tableone(main, variable.id, dataset)
}
\arguments{
\item{main}{A character for the variable names of interests.}

\item{variable.id}{A numeric/integer vector indicating the variables to compare.}

\item{dataset}{A data.frame, the annotation data which table one will be generated from.}
}
\value{
A data frame for table one visualization or export.
}
\description{
Calculate numbers and statistics for table one
}
\examples{
output <- tableone("Cancer", c(1:5), index) # Binary main variable
output <- tableone("Smoking", c(1:4, 6), index) # Categorical main variable
output <- tableone("RIN", c(2:6), index) # Continuous main variable

}
